pyproject.toml
[project]
name = "multimodal-rag"
version = "0.1.0"
description = "Entity-Aware Hierarchical Plan-Execute RAG System with pgvector and LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core LangChain & LangGraph
    "langchain-core>=0.3.72",
    "langchain>=0.3.14",
    "langgraph>=0.3.6",
    "langsmith>=0.2.9",
    # LLM Providers
    "langchain-openai>=0.3.1",
    "langchain-ollama>=0.3.1",
    "langchain-community>=0.3.14",
    # Vector Store & Database
    "langchain-postgres>=0.0.14",
    "pgvector>=0.2.5,<0.4",
    "psycopg2-binary>=2.9.10",
    # Data Processing
    "pandas>=2.3.1",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.8.1",
    # Environment & Configuration
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    # Development UI
    "streamlit>=1.41.0",
    # Utils
    "numpy>=2.0.2",
    "typing-extensions>=4.12.2",
    "psycopg[binary]>=3.2.9",
    "ipykernel>=6.30.0",
    "langchain-anthropic>=0.3.17",
    "aiosqlite>=0.21.0",
    "networkx>=3.5",
    "kiwipiepy>=0.21.0",
    "rich>=14.1.0",
    "matplotlib>=3.10.5",
    "seaborn>=0.13.2",
    "tavily-python>=0.6.0",
    "psutil>=7.0.0",
    # Async Database
    "asyncpg>=0.29.0",
    # Progress bars
    "tqdm>=4.66.0",
    # Testing
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "nest-asyncio>=1.6.0",
    "spacy>=3.8.7",
    "en-core-web-sm",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz" }
